+++ document_merge_service/api/engines.py
@@ -1,7 +1,7 @@
 import zipfile
 
 from docx import Document
-from docxtpl import DocxTemplate
+from docxtpl import DocxTemplate, Listing
 from mailmerge import MailMerge
 from rest_framework import exceptions
 
@@ -19,11 +19,28 @@ class DocxValidator:
         self.template.seek(0)
 
 
+def walk_nested(data, fn):
+    """Apply fn to every primitive value in nested data."""
+    if isinstance(data, dict):
+        return {key: walk_nested(value, fn) for (key, value) in data.items()}
+    if isinstance(data, list):
+        return [walk_nested(value, fn) for value in data]
+    return fn(data)
+
+
 class DocxTemplateEngine(DocxValidator):
     def __init__(self, template):
         self.template = template
 
     def merge(self, data, buf):
+        def wrap(value):
+            """Wrap value for multiline support."""
+            if value is None:
+                return None
+            return Listing(value)
+
+        data = walk_nested(data, wrap)
+
         doc = DocxTemplate(self.template)
 
         doc.render(data, get_jinja_env())
