+++ document_merge_service/api/tests/snapshots/snap_test_template.py
@@ -13,9 +13,7 @@ snapshots[
   <w:p>
     <w:pPr>
       <w:pStyle w:val="Normal"/>
-      <w:rPr>
-        <w:rFonts w:ascii="DejaVu Sans" w:hAnsi="DejaVu Sans"/>
-      </w:rPr>
+      <w:rPr/>
     </w:pPr>
     <w:r>
       <w:rPr>
@@ -51,9 +49,7 @@ snapshots[
   <w:p>
     <w:pPr>
       <w:pStyle w:val="Normal"/>
-      <w:rPr>
-        <w:rFonts w:ascii="DejaVu Sans" w:hAnsi="DejaVu Sans"/>
-      </w:rPr>
+      <w:rPr/>
     </w:pPr>
     <w:r>
       <w:rPr>
@@ -119,15 +115,9 @@ snapshots[
       <w:rPr>
         <w:rFonts w:ascii="DejaVu Sans" w:hAnsi="DejaVu Sans"/>
       </w:rPr>
-      <w:t xml:space="preserve"/>
-    </w:r>
-    <w:r>
-      <w:t xml:space="preserve">This is</w:t>
+      <w:t>This is</w:t>
       <w:br/>
       <w:t xml:space="preserve">a test.</w:t>
-    </w:r>
-    <w:r>
-      <w:t xml:space="preserve"/>
     </w:r>
   </w:p>
   <w:p>
@@ -180,39 +170,6 @@ snapshots[
     <w:formProt w:val="false"/>
     <w:textDirection w:val="lrTb"/>
     <w:docGrid w:type="default" w:linePitch="240" w:charSpace="0"/>
-  </w:sectPr>
-</w:body>
-"""
-
-snapshots[
-    "test_template_merge_docx[TestNameMailMerge-docx-mailmerge-template__template1] 1"
-] = """<w:body xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main" xmlns:wpc="http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:m="http://schemas.openxmlformats.org/officeDocument/2006/math" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:wp14="http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing" xmlns:wp="http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing" xmlns:w10="urn:schemas-microsoft-com:office:word" xmlns:w14="http://schemas.microsoft.com/office/word/2010/wordml" xmlns:w15="http://schemas.microsoft.com/office/word/2012/wordml" xmlns:wpg="http://schemas.microsoft.com/office/word/2010/wordprocessingGroup" xmlns:wpi="http://schemas.microsoft.com/office/word/2010/wordprocessingInk" xmlns:wne="http://schemas.microsoft.com/office/word/2006/wordml" xmlns:wps="http://schemas.microsoft.com/office/word/2010/wordprocessingShape">
-  <w:p w:rsidR="0083709B" w:rsidRPr="0083709B" w:rsidRDefault="0083709B">
-    <w:pPr>
-      <w:rPr>
-        <w:lang w:val="en-US"/>
-      </w:rPr>
-    </w:pPr>
-    <w:r>
-      <w:rPr>
-        <w:lang w:val="en-US"/>
-      </w:rPr>
-      <w:t xml:space="preserve">Test: </w:t>
-    </w:r>
-    <w:r>
-      <w:rPr>
-        <w:lang w:val="en-US"/>
-      </w:rPr>
-      <w:t>Test input</w:t>
-    </w:r>
-    <w:bookmarkStart w:id="0" w:name="_GoBack"/>
-    <w:bookmarkEnd w:id="0"/>
-  </w:p>
-  <w:sectPr w:rsidR="0083709B" w:rsidRPr="0083709B">
-    <w:pgSz w:w="11906" w:h="16838"/>
-    <w:pgMar w:top="1417" w:right="1417" w:bottom="1134" w:left="1417" w:header="708" w:footer="708" w:gutter="0"/>
-    <w:cols w:space="708"/>
-    <w:docGrid w:linePitch="360"/>
   </w:sectPr>
 </w:body>
 """
@@ -569,6 +526,39 @@ snapshots[
     <w:formProt w:val="false"/>
     <w:textDirection w:val="lrTb"/>
     <w:docGrid w:type="default" w:linePitch="240" w:charSpace="0"/>
+  </w:sectPr>
+</w:body>
+"""
+
+snapshots[
+    "test_template_merge_docx[TestNameMailMerge-docx-mailmerge-template__template1] 1"
+] = """<w:body xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main" xmlns:wpc="http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:m="http://schemas.openxmlformats.org/officeDocument/2006/math" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:wp14="http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing" xmlns:wp="http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing" xmlns:w10="urn:schemas-microsoft-com:office:word" xmlns:w14="http://schemas.microsoft.com/office/word/2010/wordml" xmlns:w15="http://schemas.microsoft.com/office/word/2012/wordml" xmlns:wpg="http://schemas.microsoft.com/office/word/2010/wordprocessingGroup" xmlns:wpi="http://schemas.microsoft.com/office/word/2010/wordprocessingInk" xmlns:wne="http://schemas.microsoft.com/office/word/2006/wordml" xmlns:wps="http://schemas.microsoft.com/office/word/2010/wordprocessingShape">
+  <w:p w:rsidR="0083709B" w:rsidRPr="0083709B" w:rsidRDefault="0083709B">
+    <w:pPr>
+      <w:rPr>
+        <w:lang w:val="en-US"/>
+      </w:rPr>
+    </w:pPr>
+    <w:r>
+      <w:rPr>
+        <w:lang w:val="en-US"/>
+      </w:rPr>
+      <w:t xml:space="preserve">Test: </w:t>
+    </w:r>
+    <w:r>
+      <w:rPr>
+        <w:lang w:val="en-US"/>
+      </w:rPr>
+      <w:t>Test input</w:t>
+    </w:r>
+    <w:bookmarkStart w:id="0" w:name="_GoBack"/>
+    <w:bookmarkEnd w:id="0"/>
+  </w:p>
+  <w:sectPr w:rsidR="0083709B" w:rsidRPr="0083709B">
+    <w:pgSz w:w="11906" w:h="16838"/>
+    <w:pgMar w:top="1417" w:right="1417" w:bottom="1134" w:left="1417" w:header="708" w:footer="708" w:gutter="0"/>
+    <w:cols w:space="708"/>
+    <w:docGrid w:linePitch="360"/>
   </w:sectPr>
 </w:body>
 """
