+++ document_merge_service/api/tests/test_template.py
@@ -8,7 +8,7 @@ from lxml import etree
 from rest_framework import status
 
 from .. import models
-from ..views import walk_nested
+from ..engines import walk_nested
 from .data import django_file
 
 
@@ -248,7 +248,7 @@ def test_template_merge_jinja_filters_docx(db, client, template, snapshot, setti
             "test_datetime": "1984-09-15 23:23",
             "test_datetime2": "23:23-1984-09-15",
             "test_none": None,
-            "test_richtext": {"nested": "This is\na test."},
+            "test_nested": {"multiline": "This is\na test."},
         }
     }
